re-factor image resampling functions into new resample module
